Converted to extension methods.
